JIN-22 textonly tests fail on android 2.2 and 2.3
Seems like the problem is that the jquery .val() on a textarea is failing.
Or perhaps rather, if you set via .html() and retreive via .val() and
visa versa there are problems.  Only on early android browsers though.

So although it is a jquery bug, it seems like there is no downside apart from slightly
increased code size to using .html() for our code.